move src/app/components dir to src/components
