+++ src/app/about/page.tsx
@@ -1,15 +1,15 @@
-'use client';
+"use client";
 
 import React, { FC } from "react";
 import { NextUIProvider } from "@nextui-org/react";
-import Layout from "../components/Layout";
-import AboutPage from "../components/AboutPage";
+import Layout from "../../components/Layout";
+import AboutPage from "../../components/AboutPage";
 
 const About: FC = () => {
   return (
     <NextUIProvider>
       <Layout>
-      <title>About - Clean and Green Philly</title>
+        <title>About - Clean and Green Philly</title>
         <AboutPage />
       </Layout>
     </NextUIProvider>
