+++ src/app/legal-disclaimer/page.tsx
@@ -1,15 +1,15 @@
-'use client';
+"use client";
 
 import React, { FC } from "react";
 import { NextUIProvider } from "@nextui-org/react";
-import Layout from "../components/Layout";
-import LegalDisclaimerPage from "../components/LegalDisclaimerPage";
+import Layout from "../../components/Layout";
+import LegalDisclaimerPage from "../../components/LegalDisclaimerPage";
 
 const LegalDisclaimer: FC = () => {
   return (
     <NextUIProvider>
       <Layout>
-      <title>Legal Disclaimer - Clean and Green Philly</title>
+        <title>Legal Disclaimer - Clean and Green Philly</title>
         <LegalDisclaimerPage />
       </Layout>
     </NextUIProvider>
