+++ src/app/methodology/page.tsx
@@ -2,14 +2,14 @@
 
 import React, { FC } from "react";
 import { NextUIProvider } from "@nextui-org/react";
-import Layout from "../components/Layout";
-import MethodologyPage from "../components/MethodologyPage";
+import Layout from "../../components/Layout";
+import MethodologyPage from "../../components/MethodologyPage";
 
 const Methodology: FC = () => {
   return (
     <NextUIProvider>
       <Layout>
-      <title>Methodology - Clean and Green Philly</title>
+        <title>Methodology - Clean and Green Philly</title>
         <MethodologyPage />
       </Layout>
     </NextUIProvider>
