+++ src/app/request-removal/page.tsx
@@ -2,14 +2,14 @@
 
 import React, { FC } from "react";
 import { NextUIProvider } from "@nextui-org/react";
-import Layout from "../components/Layout";
-import RequestRemovalPage from "../components/RequestRemovalPage";
+import Layout from "../../components/Layout";
+import RequestRemovalPage from "../../components/RequestRemovalPage";
 
 const RequestRemoval: FC = () => {
   return (
     <NextUIProvider>
       <Layout>
-      <title>Request Removal - Clean and Green Philly</title>
+        <title>Request Removal - Clean and Green Philly</title>
         <RequestRemovalPage />
       </Layout>
     </NextUIProvider>
