+++ src/app/transform-property/page.tsx
@@ -2,14 +2,14 @@
 
 import React, { FC } from "react";
 import { NextUIProvider } from "@nextui-org/react";
-import Layout from "../components/Layout";
-import TransformPropertyPage from "../components/TransformPropertyPage";
+import Layout from "../../components/Layout";
+import TransformPropertyPage from "../../components/TransformPropertyPage";
 
 const TransformProperty: FC = () => {
   return (
     <NextUIProvider>
       <Layout>
-      <title>Transform a Property - Clean and Green Philly</title>
+        <title>Transform a Property - Clean and Green Philly</title>
         <TransformPropertyPage />
       </Layout>
     </NextUIProvider>
