+++ src/components/LandingPageImage.tsx
