+++ src/components/MethodologyPage.tsx
