+++ src/components/PropertyCard.tsx
