+++ src/components/PropertyDetailSection.tsx
