+++ src/components/SidePanelControlBar.tsx
