+++ src/components/SinglePropertyDetail.tsx
