+++ src/components/SinglePropertyInfoCard.tsx
