+++ src/components/ZoomModal.tsx
